updated to work on sbt versions 0.10.1 as well as 0.7.5
Updated library dependency for gson to 1.7.1 because 1.7-SNAPSHOT is no longer
available.  Moved all 0.7.5 project code to project-0.7.5.

Implemented new project code for 0.10.1 that uses the onLoad hook to update
the javaOptions classpath at load time. That pile of code replaces the
"override def fork =..." line originally used for forking.  Not sure if this
problem is due to caliper or sbt...